[ADD] Backend Info DB Claims
Se configuro apartados como el mensaje de error con contraseñas incorrectas y se agrego aparte del token y adentro del token

+++ backend/src/main/java/gaseco/backend/Auth/RegisterRequest.java
@@ -32,5 +32,5 @@ public class RegisterRequest {
     String cvesucursal;
     String cvealmacen;
     LocalDate fhVencimineto;
-    BigDecimal intentosPassword;
+    int intentosPassword;
 }
